Update la tipul de date

+++ HeapSort/heapsort.cpp
@@ -27,13 +27,13 @@ Mica introducere:
 */
 
 std::vector<double> v;
-int n;
+unsigned long long n;
 
 void citire(char *filename) {
     std::ifstream fin(filename);
     fin >> n;
     double x;
-    for(int i = 0; i < n; ++i) {
+    for(unsigned long long i = 0; i < n; ++i) {
         fin >> x;
         v.push_back(x);
     }
@@ -41,7 +41,7 @@ void citire(char *filename) {
 }
 
 
-void heapify(std::vector<double> &v, int n, int root_pos) {
+void heapify(std::vector<double> &v, unsigned long long n, int root_pos) {
     // calculam indici celor 2 fii
     int left = 2 * root_pos + 1;
     int right = 2 * root_pos + 2;
@@ -64,7 +64,7 @@ void heapify(std::vector<double> &v, int n, int root_pos) {
     }
 }
 
-void heapsort(std::vector<double> &v, int n) {
+void heapsort(std::vector<double> &v, unsigned long long n) {
     // transformam arborele binar intr-un max-heap
     // pornesc de la radacina fiecarui subarbore si tot urc in inaltime
     for(int i = n / 2 - 1; i >= 0; --i) {
@@ -83,7 +83,7 @@ void heapsort(std::vector<double> &v, int n) {
 
 
 void afisare(std::vector<double
-> v, int n, char *filename) {
+> v, unsigned long long n, char *filename) {
     std::ofstream fout(filename);
 
     for(int i = 0; i < n; ++i) {
